Implement Basic Version of Favourites Page (#57)
* Created Favourites button in navbar

Also set up the Favourites page folder, following the convention established by the previous group. The index and css files are yet to be implemented.

* Placeholder implementation of favourites page

* Created route for favourites page

* /fix - route to favourites page

Fixed the route, was previously incorrectly set so that only unauthenticated users could access the favourites page, and authenticated users were redirected to login. Fixed to now be vice versa

* Fixed typo in App.jsx